8. Что такое валидация? [#246864]

+++ src/main/java/ru/job4j/tracker/Tracker.java
@@ -61,7 +61,7 @@ public class Tracker {
     }
 
     public boolean replace(int id, Item item) {
-        byte index = (byte) indexOf(id);
+        int index = indexOf(id);
         if (index != -1) {
             int idsKey = items[index].getId();
             items[index] = item;
@@ -72,7 +72,7 @@ public class Tracker {
     }
 
     public void delete(int id) {
-        byte index = (byte) indexOf(id);
+        int index = indexOf(id);
         if (index != -1) {
             System.arraycopy(items, index + 1, items, index, size - index - 1);
             items[size--] = null;
